pyproject.toml
[tool.poetry]
name = "short-url"
version = "0.1.0"
description = ""
authors = ["Minister944"]
readme = "README.md"
packages = [{include = "short_url"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.92.0"
uvicorn = "^0.20.0"
sqlalchemy = "^2.0.4"
alembic = "^1.9.4"
python-dotenv = "^0.21.1"
psycopg2-binary = "^2.9.5"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
